Infer LOCALEDIR and PKGDATADIR dynamically
Currently LOCALEDIR and PKGDATADIR are derived at build time as an
absolute path and added into the executable. This means that once the
executable has an install dir set, it cannot be moved.

While this is not an issue in Flatpak, it is clearly undesirable in some
portable environments that do not make use of Flatpak, such as AppImage,
Windows and macOS.

The LOCALEDIR and RESOURCES_FILE strings are replaced with a dynamic
string compiled in the main function itself, which means that it will
always be $BINDIR/../share/{cartero, locale}, where $BINDIR is the
location of the executable. Initially it should be $PREFIX/bin, but it
could be other things.

On MS Windows, it is expected that this should allow in the future to
ask the user for the install location when installing Cartero, the day
it is available for Windows.