Properly rebuild the .ui files from the .blp
On some cases Meson will forget to rebuild the .ui files from the .blp
files when recompiling the application.

It seems that the "official" template provided by blueprint-compiler
that most projects use set the output to "." rather than to the real
list of .ui files that should be generated by the compiler, which causes
the dependencies to not be checked.

Now, if you edit one single .blp file, every other .ui will be rebuilt.
While it shouldn't have to rebuild every file, at least it is something.