Enhance compatibility with Windows
This should be tested outside Windows ASAP to assert that the compiler
directive added to src/main.rs doesn't break other platforms, but the
changes made to src/main.rs should prevent a Windows command prompt from
opening when starting the application.

Also, I've fixed src/meson.build so that it builds on Windows properly
without the `&&` bug, and I documented the build instructions on
WINDOWS.md for the future. They should be moved somewhere else.