Refactor KeyValuePane so that it uses GtkListBox
Instead of being a GtkListView, replace it with a GtkListBox. The
behaviour is similar, but it makes unnecessary to do things such as
setup a factory, it allows to control the model in a more direct way,
and it allows me to avoid the gray backgrounds when the mouse hovers a
row.