Refactor access to settings
It turns out that it is possible to get multiple references to the same
Settings object, because it turns out that the application is actually
treated as a singleton by Gio, which means that I don't have to pass
references to the Settings object during instantiation, but just access
the default application and get the settings reference bound to it.

In this commit, CarteroApplication::get is added as a function that
returns the CarteroApplication instance in use. Those widgets that want
to use settings to persist stuff have been modified so that accessing
the settings is done using this global instance.