Apply sane XDG conventions and fix datafile issues
In order to make the application portable (specially on Microsoft
Windows), data directory lookup should not be hardcoded. After applying
this commit, the application will add $PWD/../share to the XDG_DATA_DIRS
path if it wasn't already present.

It should not be necessary to manually set GSETTINGS_SCHEMA_DIR when
running the application if it is not installed to a standard location.

Note that this will still assume that the gdk-pixbufs, the default
icon themes and the GtkSourceView 5 support files are installed at the
standard location. This should be the case for GNU/Linux, but may not be
the case for other operating systems like Windows or macOS. This is
something that I've recently learnt by reading the build-mingw.sh script
of the GNOME Text Editor source code.

I've modified the cargo-build.sh file as well in order to copy the
application icons as well, because otherwise some parts like the
AboutWindow may not pick the icon properly. I've also updated the
README.md to reflect the changes made recently to the build process.