Replace header and variable maps with header lists
This change is done because multiple items with the same key may exist
at the same time, so a regular Map<String, String> cannot be used, or
the duplicate information may be lost.

This commit changes a lot of structs, but the key things is that the
local data structures such as Endpoint, Request or RequestMethod are now
moved into entities.rs.