Treat client side decorations as a Rust feature
This is mainly used for the Windows and Mac versions. If the "csd"
feature (which is enabled by default) gets disabled, CarteroWindow will
extend from Gtk.ApplicationWindow so that no client side decorations are
used as long as the window title widget is not set. Otherwise, it will
continue extending Adw.ApplicationWindow and the resulting window will
have client side decorations.

This is done because, as I explain in
https://github.com/danirod/cartero/issues/25, client side decorations
have a bad integration on Windows and macOS environments currently. My
nitpicks are:

* (Critical) Client-side decorations do not support native Snap Window,
  so it is more difficult to manage the window.
* (Critical) Client-side decorations may affect the way the window
  bounds are perceived by the OS. The shadow is part of the window, so
  a screenshot of the window will always be bigger.
* (Normal) Client-side decorations do not follow macOS window decoration
  pattern. Specifically, they don't use the window traffic light and the
  close button appears in the other side of the window by default.
* (Less critical) On some configurations, the client side decorations
  may cause black borders to appear on Windows.