Add more tests to make the files more robust
Specifically, I've changed the implementation of Ord for headers so that
only the name is considered during sort. On stable sorts, this will
guarantee that an array of entries will preserve the order of items with
the same name, which is important during file serialization and requests
since positions that appear near the end of the array overwrite previous
positions.